+++ pages/_app.js
@@ -1,5 +1,5 @@
 import GlobalStyle from "../styles";
-import Layout from "@/Layout";
+import Layout from "@/Components/Layout";
 
 export default function App({ Component, pageProps }) {
   return (
